Comment delete feature: - updated with upvote orphan delete.
