+++ src/main/java/com/roadmaps/Roadmaps/modules/roadmap/service/UpvoteService.java
@@ -13,6 +13,4 @@ public interface UpvoteService {
     List<Upvote> getUpvotesByCommentId(UUID commentId);
 
     List<Upvote> upvoteByUserEmailAndCommentId(String userEmail, String roadmapId);
-
-    void deleteAllUpvoteByCommentIds(List<UUID> allCommentIds);
 }
